% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEXTFLOW.R
\name{Run}
\alias{Run}
\title{Run R6Class}
\description{
Run R6Class

Run R6Class
}
\details{
An R6 object representing a Nextflow pipeline run. The class manages the execution
of the pipeline, as well as the retrieval of results and logs, and cleaning up
after execution.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pipeline_path}}{The path to the pipeline script or project directory.}

\item{\code{run_path}}{The path to the directory where the pipeline results will be stored.}

\item{\code{name}}{An optional name for the run. Defaults to a digest of \code{run_path} if \code{NULL}.}

\item{\code{nextflow_parameters}}{A \code{data.table} holding parameters specific to Nextflow.}

\item{\code{tool_parameters}}{A \code{data.table} holding parameters specific to the pipeline tools.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Run-new}{\code{Run$new()}}
\item \href{#method-Run-execute}{\code{Run$execute()}}
\item \href{#method-Run-get_results}{\code{Run$get_results()}}
\item \href{#method-Run-get_logs}{\code{Run$get_logs()}}
\item \href{#method-Run-clean}{\code{Run$clean()}}
\item \href{#method-Run-clone}{\code{Run$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-new"></a>}}
\if{latex}{\out{\hypertarget{method-Run-new}{}}}
\subsection{Method \code{new()}}{
Initialize method of Run object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$new(pipeline_path, run_path, name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pipeline_path}}{path to the pipeline script or project directory}

\item{\code{run_path}}{path to the directory where the pipeline results will be stored}

\item{\code{name}}{an optional name for the run, defaults to a digest of run_path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-execute"></a>}}
\if{latex}{\out{\hypertarget{method-Run-execute}{}}}
\subsection{Method \code{execute()}}{
Execute method for the Nextflow pipeline
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$execute(samplesheet)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{samplesheet}}{data.table or data.frame with samplesheet information to run the pipeline on}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-Run-get_results}{}}}
\subsection{Method \code{get_results()}}{
Get the results of a pipeline run
Retrieves results from the results directory after the pipeline has been executed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$get_results()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.table containing the paths to the results files.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-get_logs"></a>}}
\if{latex}{\out{\hypertarget{method-Run-get_logs}{}}}
\subsection{Method \code{get_logs()}}{
Get the logs of a Nextflow pipeline
Extracts and returns the Nextflow logs of the pipeline run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$get_logs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.table containing extracted log information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-clean"></a>}}
\if{latex}{\out{\hypertarget{method-Run-clean}{}}}
\subsection{Method \code{clean()}}{
Clean method for cleaning up the Nextflow work directory
Executes the \verb{nextflow clean} command with optional parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$clean(clean_params = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clean_params}}{character string with additional parameters for the \verb{nextflow clean} command}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Run-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
